requirements.txt
Django>=5.1,<6.1
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
boto3>=1.35,<2.0
python-dotenv>=1.0,<2.0
djangorestframework>=3.15,<4.0
whitenoise>=6.8,<7.0
uvicorn[standard]>=0.30,<1.0
python-socketio>=5.11,<6.0
